requirements.txt
easydict
fastapi
opencv-python
Pillow
PyYAML
torch
torchvision
uvicorn
onnxruntime
python-multipart
